+++ config/run/exec.go
@@ -7,11 +7,15 @@ import (
 	"github.com/rliebz/tusk/ui"
 )
 
+const shellEnvVar = "SHELL"
+const defaultShell = "/bin/sh"
+
 // ExecCommand executes a shell command.
 func ExecCommand(command string) error {
 	ui.PrintCommand(command)
 
-	cmd := exec.Command("sh", "-c", command) // nolint: gas
+	shell := getShell()
+	cmd := exec.Command(shell, "-c", command) // nolint: gas
 	cmd.Stdin = os.Stdin
 	if ui.Verbosity > ui.VerbosityLevelSilent {
 		cmd.Stdout = os.Stdout
@@ -26,4 +30,13 @@ func ExecCommand(command string) error {
 	}
 
 	return nil
+}
+
+// getShell returns the value of the `SHELL` environment variable, or `/bin/sh`.
+func getShell() string {
+	if shell := os.Getenv(shellEnvVar); shell != "" {
+		return shell
+	}
+
+	return defaultShell
 }